Changed ID's to long. Fixed all compile time errors

+++ WebModel/src/main/java/org/kofi/creditex/web/model/CreditApplicationForm.java
@@ -6,11 +6,11 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class CreditApplicationForm {
-    private int id;
+    private long id;
     private String productName;
-    private int productId;
-    private int requestedMoney;
-    private int duration;
+    private long productId;
+    private long requestedMoney;
+    private long duration;
     private String acceptance;
     private String applicationDate;
     private String whoRejected;
