+++ WebModel/src/main/java/org/kofi/creditex/web/model/CreditForm.java
@@ -8,14 +8,14 @@ import java.util.List;
 @Data
 @Accessors(chain = true)
 public class CreditForm {
-    private int id;
+    private long id;
     private String start;
-    private int duration;
-    private int currentMainDebt;
-    private int fine;
-    private int currentMoney;
-    private int originalMainDebt;
-    private int productId;
+    private long duration;
+    private long currentMainDebt;
+    private long fine;
+    private long currentMoney;
+    private long originalMainDebt;
+    private long productId;
     private String productName;
     private List<PaymentTableForm> payments;
 }
