+++ WebModel/src/main/java/org/kofi/creditex/web/model/PaymentTableForm.java
@@ -6,9 +6,9 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class PaymentTableForm {
-    private int id;
-    private int number;
-    private int payment;
+    private long id;
+    private long number;
+    private long payment;
     private String start;
     private String end;
     private String expired;
