+++ WebModel/src/main/java/org/kofi/creditex/web/model/PriorApplicationForm.java
@@ -6,9 +6,9 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class PriorApplicationForm {
-    private int id;
+    private long id;
     private String comment;
-    private int creditId;
+    private long creditId;
     private String acceptance;
     private String date;
 }
