+++ WebModel/src/main/java/org/kofi/creditex/web/model/UserChangeDataForm.java
@@ -6,7 +6,7 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class UserChangeDataForm {
-    private int id;
+    private long id;
     private String first;
     private String last;
     private String patronymic;
@@ -14,5 +14,5 @@ public class UserChangeDataForm {
     private int passportNumber;
     private String workName;
     private String workPosition;
-    private int workIncome;
+    private long workIncome;
 }
