+++ repository/src/main/java/org/kofi/creditex/repository/AuthoritiesRepository.java
@@ -4,6 +4,6 @@ import org.kofi.creditex.model.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
-public interface AuthoritiesRepository extends JpaRepository<Authority, Integer>, QueryDslPredicateExecutor<Authority> {
+public interface AuthoritiesRepository extends JpaRepository<Authority, Long>, QueryDslPredicateExecutor<Authority> {
     Authority findByAuthority(String authority);
 }
