+++ repository/src/main/java/org/kofi/creditex/repository/OperationRepository.java
@@ -4,5 +4,5 @@ import org.kofi.creditex.model.Operation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
-public interface OperationRepository extends JpaRepository<Operation, Integer>, QueryDslPredicateExecutor<Operation> {
+public interface OperationRepository extends JpaRepository<Operation, Long>, QueryDslPredicateExecutor<Operation> {
 }
