+++ repository/src/main/java/org/kofi/creditex/repository/PaymentRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 import java.sql.Date;
 import java.util.List;
 
-public interface PaymentRepository extends JpaRepository<Payment, Integer>, QueryDslPredicateExecutor<Payment> {
-    List<Payment> findByCredit_Id(int id);
+public interface PaymentRepository extends JpaRepository<Payment, Long>, QueryDslPredicateExecutor<Payment> {
+    List<Payment> findByCredit_Id(long id);
     List<Payment> findByCredit_RunningAndPaymentClosedAndPaymentEndLessThan(boolean creditRunning,boolean closedPayment, Date dateNow);
     List<Payment> findByCredit_RunningAndPaymentExpiredAndPaymentExpiredProcessed(boolean creditRunning, boolean paymentExpired, boolean paymentExpiredProcessed);
 }
