+++ repository/src/main/java/org/kofi/creditex/repository/ProductRepository.java
@@ -4,5 +4,5 @@ import org.kofi.creditex.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
-public interface ProductRepository extends JpaRepository<Product, Integer>, QueryDslPredicateExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Long>, QueryDslPredicateExecutor<Product> {
 }
