+++ repository/src/main/java/org/kofi/creditex/repository/ProlongationApplicationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
 import java.util.List;
 
-public interface ProlongationApplicationRepository extends JpaRepository<ProlongationApplication, Integer>, QueryDslPredicateExecutor<ProlongationApplication> {
+public interface ProlongationApplicationRepository extends JpaRepository<ProlongationApplication, Long>, QueryDslPredicateExecutor<ProlongationApplication> {
     List<ProlongationApplication> findByClient_Username(String username);
 }
