+++ repository/src/main/java/org/kofi/creditex/repository/UserDataRepository.java
@@ -4,6 +4,6 @@ import org.kofi.creditex.model.UserData;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
-public interface UserDataRepository extends JpaRepository<UserData, Integer>, QueryDslPredicateExecutor<UserData> {
+public interface UserDataRepository extends JpaRepository<UserData, Long>, QueryDslPredicateExecutor<UserData> {
     UserData findByUserUsername(String name);
 }
